[#216] Rename dialogue files
