+++ Assets/Scripts/UI/Dialogue/DialogueController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 1cd4b9417c80a444bb9dfddf90a09016
+guid: 8ea77156d8c39a94eb51b7308ca1819a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
