+++ Assets/Scripts/UI/Dialogue/DialogueInteraction.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e7ab563f6092a464dbf57c9ff7a27d3b
+guid: 1cd4b9417c80a444bb9dfddf90a09016
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
