--- Assets/Scripts/UI/Dialogue/DialogueManagerNew.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8ea77156d8c39a94eb51b7308ca1819a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
