+++ Assets/Scripts/UI/Dialogue/DialogueTrigger.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 05571a748d7b98f41a31b6ffcd22732c
+guid: 6c570f723c37fbe4989a81133b04b079
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
