--- Assets/Scripts/UI/Dialogue/DialogueTriggerNew.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6c570f723c37fbe4989a81133b04b079
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
